Split repository.ts: extract member discovery, digests, access requests (3/N)
Third slice of the repository.ts carve-up (audit L14), continuing #804 and
#807. Moves three more domains out behind the same re-export facade:

- repository/memberDiscovery.ts — published member interests (#634) plus the
  helper handoff behind set_helper_availability / find_helper (#729). Kept as
  ONE module because find_helper's match query is deliberately modelled on
  searchMemberInterests' shape and the doc comments cross-reference it as
  "above"; splitting them would make those comments lie.
- repository/contextDigests.ts — the nightly offline digest-builder output (#51).
- repository/accessRequests.ts — the gated-mode pending access-request queue.

Also introduces repository/shared.ts for cross-cutting internals, holding
KNOWLEDGE_SEARCH_RELEVANCE_THRESHOLD. A domain module needed that floor, and
importing it back from repository.ts would be a cycle (repository.ts
re-exports the domain modules) — the exact circularity the constant's own
comment says it was placed in repository.ts to avoid. repository.ts imports it
explicitly for its own knowledge queries, since `export *` does not bind names
into the re-exporting module's scope.

Pure code motion: all 27 moved symbols verified byte-identical against
origin/main mechanically, and every import site plus tests/repository.test.ts
are untouched. repository.ts is now 6,550 lines, down from 7,137 at the start
of the campaign.

Screening done before cutting, per the lessons of the prior two PRs: each
candidate section was checked for (a) calls out to helpers defined elsewhere in
the file, (b) reverse references from the remaining body — all four outside
references turned out to be doc-comment mentions, not code — and (c) tests that
slice repository.ts as source text between `// --- ` banners. Typecheck also
caught one bare-constant dependency the call-pattern screen missed, which is
what motivated shared.ts.
