+++ src/storage/repository.ts
@@ -31,7 +31,14 @@ import { pageKeyOf } from '../context/docsIngest.js';
  */
 export * from './repository/preferences.js';
 export * from './repository/memberNotes.js';
+export * from './repository/shared.js';
 export * from './repository/devTeamWatches.js';
+// `export *` does not bind names into this module's own scope, and the
+// knowledge queries below use this floor directly — so import it explicitly.
+import { KNOWLEDGE_SEARCH_RELEVANCE_THRESHOLD } from './repository/shared.js';
+export * from './repository/accessRequests.js';
+export * from './repository/contextDigests.js';
+export * from './repository/memberDiscovery.js';
 
 // `export *` re-exports for CALLERS but does not bind the names in this
 // module's own scope, so anything still living here that uses an extracted
@@ -750,40 +757,6 @@ export async function saveKnowledge(input: {
 
   return { id: Number(rows[0].id), similarEntry };
 }
-
-/**
- * Relevance floor for `knowledge_search` hits, in cosine similarity
- * (`1 - (embedding <=> query)`, same units as `searchKnowledge`'s returned
- * `similarity`). This is a *relevance* floor ("is this topically usable at
- * all"), not a *duplicate* floor like `QUESTION_CLUSTER_SIMILARITY_THRESHOLD`
- * below (0.85, "is this the same question") — it is deliberately much lower.
- *
- * The value is a function of the current embedding model
- * (`config.db.embeddingModel`, currently Xenova/all-MiniLM-L6-v2) and query
- * distribution, not a universal constant. It was derived empirically against
- * `tests/fixtures/knowledgeEval.json` (see the `negativeQueries` case in
- * knowledgeEval.test.ts): with this model, unambiguously off-topic queries
- * (e.g. "what's the best coffee place near the venue") score ~0.15-0.22
- * against every fixture entry, and a topically-adjacent near-miss (asking how
- * long admin applications take to hear back — same topic as "Requesting admin
- * role", but a question that entry doesn't answer) tops out at ~0.33, while
- * all but a couple of the weakest genuine paraphrase matches score 0.36+. A
- * small minority of very loosely-worded genuine matches score below this
- * floor too (e.g. "what are the guidelines for behaving in this server" vs.
- * the actual "Discord server rules" entry, ~0.30) — that's an intentional
- * precision-over-recall trade-off: this feature exists specifically so a
- * low-confidence hit results in "no confident match" (which the system
- * prompt turns into an honest hedge) rather than a shaky answer stated as
- * fact. If `EMBEDDING_MODEL` ever changes, this constant must be re-derived
- * the same way — a model swap will otherwise silently degrade filtering with
- * no test failure.
- *
- * Defined here (not in agent/tools.ts, which re-exports it for
- * `knowledge_search`'s own filtering) so `knowledgeCoversTopic` below — the
- * issue #102 candidate dedup guard — can share the exact same floor without
- * agent/tools.ts and storage/repository.ts importing each other.
- */
-export const KNOWLEDGE_SEARCH_RELEVANCE_THRESHOLD = 0.35;
 
 /**
  * Mark unresolved `knowledge_gaps` rows resolved when `embedding` (the
@@ -3035,179 +3008,6 @@ export async function recordAdminAction(
   );
 }
 
-// --- Access requests (gated-mode pending queue) -----------------------------
-
-/**
- * Record that a gated guest addressed the bot. Identity + counts only — the
- * caller must never pass message content. Upserts so repeat pings from the
- * same user dedup into one row instead of growing unbounded.
- *
- * Returns whether this call created a FRESH row (`inserted`), via Postgres's
- * own `xmax = 0` trick on `RETURNING` — distinguishing "first insert" from
- * "repeat upsert" needs no extra query or column, just reading what the
- * upsert already tells us (issue #480). This is the debounce signal
- * `notifyAccessRequest`'s first-time-only real-time alert relies on: a
- * repeat ping from the same still-pending guest returns `inserted: false` and
- * must not notify again.
- *
- * Also returns `firstRequestedAt` (issue #591) — read off the same row via
- * the same `RETURNING` clause, so surfacing it to the returning-guest wait
- * clause in the gated notice costs zero new queries and zero new columns.
- */
-export async function recordAccessRequest(input: {
-  platform: Platform;
-  userId: string;
-  userName?: string;
-}): Promise<{ inserted: boolean; firstRequestedAt: Date }> {
-  const { rows } = await pool.query(
-    `INSERT INTO access_requests (platform, user_id, user_name)
-     VALUES ($1,$2,$3)
-     ON CONFLICT (platform, user_id) DO UPDATE
-       SET last_requested_at = now(),
-           request_count = access_requests.request_count + 1,
-           user_name = COALESCE(EXCLUDED.user_name, access_requests.user_name)
-     RETURNING (xmax = 0) AS inserted, first_requested_at`,
-    [input.platform, input.userId, input.userName ?? null],
-  );
-  return { inserted: rows[0]?.inserted === true, firstRequestedAt: rows[0].first_requested_at };
-}
-
-export interface AccessRequest {
-  platform: Platform;
-  userId: string;
-  userName: string | null;
-  firstRequestedAt: Date;
-  lastRequestedAt: Date;
-  requestCount: number;
-}
-
-export async function listAccessRequests(limit = 50): Promise<AccessRequest[]> {
-  const { rows } = await pool.query(
-    `SELECT platform, user_id, user_name, first_requested_at, last_requested_at, request_count
-       FROM access_requests
-      ORDER BY last_requested_at DESC
-      LIMIT $1`,
-    [limit],
-  );
-  return rows.map((r) => ({
-    platform: r.platform,
-    userId: r.user_id,
-    userName: r.user_name,
-    firstRequestedAt: r.first_requested_at,
-    lastRequestedAt: r.last_requested_at,
-    requestCount: Number(r.request_count),
-  }));
-}
-
-/** Clear a resolved access request (e.g. after add_member succeeds for that user). */
-export async function clearAccessRequest(platform: Platform, userId: string): Promise<void> {
-  await pool.query(`DELETE FROM access_requests WHERE platform = $1 AND user_id = $2`, [platform, userId]);
-}
-
-/**
- * Exact pending-guest count — a dedicated `COUNT(*)` rather than
- * `(await listAccessRequests()).length`, which would silently understate a
- * backlog past that function's `limit` (default 50) cap.
- */
-export async function countAccessRequests(): Promise<number> {
-  const { rows } = await pool.query(`SELECT count(*) AS n FROM access_requests`);
-  return Number(rows[0].n);
-}
-
-/**
- * Whole-day age of the oldest still-pending access request — the same
- * `MIN(first_requested_at)` oldest-age mechanic issue #450 applies to
- * reports/suggestions, applied here to `access_requests` (issue #515).
- * `first_requested_at` is set once at insert and never updated
- * (`recordAccessRequest`), and `clearAccessRequest` deletes the row on
- * `add_member`, so by construction every remaining row is unresolved
- * backlog and `MIN` over an empty table is `null`, never `0` — returned
- * as-is rather than coerced, so an admin/digest reader can never mistake
- * "no pending requests" for "a request that just arrived".
- */
-export async function oldestAccessRequestAgeDays(): Promise<number | null> {
-  const { rows } = await pool.query(
-    `SELECT EXTRACT(DAY FROM now() - MIN(first_requested_at))::int AS age_days FROM access_requests`,
-  );
-  const ageDays = rows[0]?.age_days;
-  return ageDays === null || ageDays === undefined ? null : Number(ageDays);
-}
-
-// --- Context digests (offline builder output, issue #51) ---------------------
-
-export interface ContextDigest {
-  id: number;
-  periodStart: Date;
-  periodEnd: Date;
-  platform: string | null;
-  topic: string;
-  summary: string;
-  exampleRefs: number[];
-  distinctUsers: number;
-  questionCount: number;
-  createdAt: Date;
-}
-
-export async function insertContextDigest(input: {
-  periodStart: Date;
-  periodEnd: Date;
-  platform?: string;
-  topic: string;
-  summary: string;
-  exampleRefs: number[];
-  distinctUsers: number;
-  questionCount: number;
-}): Promise<number> {
-  const { rows } = await pool.query(
-    `INSERT INTO context_digests
-       (period_start, period_end, platform, topic, summary, example_refs, distinct_users, question_count)
-     VALUES ($1,$2,$3,$4,$5,$6,$7,$8) RETURNING id`,
-    [
-      input.periodStart,
-      input.periodEnd,
-      input.platform ?? null,
-      input.topic,
-      input.summary,
-      input.exampleRefs,
-      input.distinctUsers,
-      input.questionCount,
-    ],
-  );
-  return Number(rows[0].id);
-}
-
-export async function listContextDigests(days = 30, limit = 20): Promise<ContextDigest[]> {
-  const clampedDays = Math.min(Math.max(Math.trunc(days) || 30, 1), 365);
-  const clampedLimit = Math.min(Math.max(Math.trunc(limit) || 20, 1), 100);
-  const { rows } = await pool.query(
-    `SELECT id, period_start, period_end, platform, topic, summary, example_refs,
-            distinct_users, question_count, created_at
-       FROM context_digests
-      WHERE created_at > now() - $1::interval
-      ORDER BY created_at DESC, question_count DESC
-      LIMIT $2`,
-    [`${clampedDays} days`, clampedLimit],
-  );
-  return rows.map((r) => ({
-    id: Number(r.id),
-    periodStart: r.period_start,
-    periodEnd: r.period_end,
-    platform: r.platform,
-    topic: r.topic,
-    summary: r.summary,
-    exampleRefs: (r.example_refs as unknown[]).map(Number),
-    distinctUsers: Number(r.distinct_users),
-    questionCount: Number(r.question_count),
-    createdAt: r.created_at,
-  }));
-}
-
-/** When the builder last produced anything — backs the ~daily freshness guard. */
-export async function latestContextDigestAt(): Promise<Date | null> {
-  const { rows } = await pool.query(`SELECT max(created_at) AS at FROM context_digests`);
-  return rows[0]?.at ?? null;
-}
-
 // --- Knowledge candidates (issue #102, the knowledge_candidates half of #51
 // its adversarial review deferred) --------------------------------------------
 
@@ -3811,274 +3611,6 @@ export async function resolveSuggestion(
   return rows[0]
     ? { platform: rows[0].platform as Platform, userId: rows[0].user_id, content: rows[0].content }
     : null;
-}
-
-// --- Member interests (member-to-member discovery, issue #634) --------------
-//
-// Self-declared-member-table pattern member_projects below reuses: opt-in,
-// self-scoped, embedded, purge/roster-leave-cleaned data. One fuzzy free-text
-// blob per identity rather than discrete named artifacts, hence the plain
-// (platform, user_id) primary key and upsert-only-in-place semantics — no
-// per-member cap or rate limit is needed since a member can only ever have
-// the single row their own writes replace.
-
-/** Hard cap on published interest text length (self-declared free text), enforced server-side. */
-export const MEMBER_INTERESTS_MAX_CHARS = 300;
-/** who_is_into's result cap. */
-export const WHO_IS_INTO_LIMIT = 5;
-
-export interface MemberInterestSearchHit {
-  platform: Platform;
-  userId: string;
-  interests: string;
-  similarity: number;
-}
-
-/**
- * Self-scoped upsert/clear of the caller's OWN published interests row.
- * Passing the literal string 'clear' (case-insensitive, whitespace-trimmed)
- * deletes the row instead of writing one — the exact `text | 'clear'`
- * interface the tool description exposes. Never CONFIRM-gated: instantly
- * reversible, same shape as set_response_style/set_language_preference.
- */
-export async function setMemberInterests(
-  platform: Platform,
-  userId: string,
-  interests: string,
-): Promise<{ cleared: boolean }> {
-  if (interests.trim().toLowerCase() === 'clear') {
-    await pool.query(`DELETE FROM member_interests WHERE platform = $1 AND user_id = $2`, [platform, userId]);
-    return { cleared: true };
-  }
-  const text = interests.slice(0, MEMBER_INTERESTS_MAX_CHARS);
-  let embedding: number[] | null = null;
-  try {
-    embedding = await embed(text);
-  } catch (err) {
-    logger.warn({ err }, 'Embedding failed for member interests');
-  }
-  await pool.query(
-    `INSERT INTO member_interests (platform, user_id, interests, embedding, updated_at)
-     VALUES ($1, $2, $3, $4, now())
-     ON CONFLICT (platform, user_id) DO UPDATE SET
-       interests = EXCLUDED.interests, embedding = EXCLUDED.embedding, updated_at = now()`,
-    [platform, userId, text, embedding ? pgvector.toSql(embedding) : null],
-  );
-  return { cleared: false };
-}
-
-/**
- * Embedding-similarity search over published interests ONLY (never
- * `interactions` — SECURITY: issue #634 AC #4, interests are never inferred
- * from chat content). A caller with no row of their own can still search;
- * absent/cleared rows never appear since they simply don't exist in the table.
- */
-export async function searchMemberInterests(
-  query: string,
-  limit = WHO_IS_INTO_LIMIT,
-): Promise<MemberInterestSearchHit[]> {
-  const clampedLimit = Math.min(Math.max(Math.trunc(limit) || WHO_IS_INTO_LIMIT, 1), 50);
-  let queryVec: number[];
-  try {
-    queryVec = await embed(query);
-  } catch (err) {
-    logger.warn({ err }, 'Embedding query failed; skipping member interest search');
-    return [];
-  }
-  const { rows } = await pool.query(
-    `SELECT platform, user_id, interests, 1 - (embedding <=> $1) AS similarity
-       FROM member_interests
-      WHERE embedding IS NOT NULL
-      ORDER BY embedding <=> $1
-      LIMIT $2`,
-    [pgvector.toSql(queryVec), clampedLimit],
-  );
-  return rows.map((r) => ({
-    platform: r.platform as Platform,
-    userId: r.user_id,
-    interests: r.interests as string,
-    similarity: Number(r.similarity),
-  }));
-}
-
-/**
- * Batched lookup of published `member_interests` text for a set of owners,
- * keyed by `"platform:userId"` — used by list_projects' cross-reference
- * (issue #718) to show a project owner's published interests without an
- * N+1 query per rendered row. One round trip for the whole result set via
- * `unnest($1, $2)` zipping the parallel platform/userId arrays into rows,
- * matched against the composite `(platform, user_id)` key — SECURITY:
- * issue #718 AC #7, this can only ever return rows for owners in the input
- * set, never a full-table read. An owner with no published interests (or
- * who cleared them) simply has no entry in the returned Map.
- */
-export async function getPublishedInterestsForOwners(
-  owners: ReadonlyArray<{ platform: Platform; userId: string }>,
-): Promise<Map<string, string>> {
-  const result = new Map<string, string>();
-  if (owners.length === 0) return result;
-  const { rows } = await pool.query(
-    `SELECT platform, user_id, interests
-       FROM member_interests
-      WHERE (platform, user_id) IN (SELECT * FROM unnest($1::text[], $2::text[]))`,
-    [owners.map((o) => o.platform), owners.map((o) => o.userId)],
-  );
-  for (const r of rows) {
-    result.set(`${r.platform}:${r.user_id}`, r.interests as string);
-  }
-  return result;
-}
-
-// --- Helper handoff (set_helper_availability / find_helper, issue #729) -----
-//
-// The active-side consumer of member_interests: willing_to_help rides the
-// same row set_my_interests/who_is_into already publish/search, so a helper
-// must have a published interests row before they can opt in. Matching reuses
-// searchMemberInterests' exact query shape (embedding-similarity, embedding
-// IS NOT NULL) filtered to willing_to_help = true and excluding the
-// requester's own row. helper_notifications is a separate append-only log,
-// never edited in place, backing two independent DB-backed rolling-window
-// caps (never in-memory counters, so both survive a process restart).
-
-/** Hard cap on a find_helper topic's length, mirroring rate_answer's comment cap. */
-export const FIND_HELPER_TOPIC_MAX_CHARS = 200;
-/** How many top-ranked candidates find_helper will walk before giving up — bounded above the two rate caps below so a scan isn't starved by a few maxed-out helpers. */
-export const FIND_HELPER_CANDIDATE_SCAN_LIMIT = 10;
-/** Per-helper cap on notifications received in a rolling 7 days — the "unsolicited pings" guardrail. */
-export const FIND_HELPER_WEEKLY_LIMIT_PER_HELPER = 3;
-/** Per-requester cap on find_helper calls in a rolling 24h — prevents looping over topics to exhaust many helpers' weekly quotas. */
-export const FIND_HELPER_REQUESTER_DAILY_LIMIT = 3;
-/**
- * Minimum cosine similarity for a willing_to_help row to count as a genuine
- * match — same floor and same embedding model as
- * KNOWLEDGE_SEARCH_RELEVANCE_THRESHOLD. Without this, findHelperCandidates
- * would always return the nearest-ranked willing helper regardless of actual
- * relevance (searchMemberInterests/who_is_into has the same no-floor shape,
- * but it only ever surfaces a list for the REQUESTER to judge; find_helper
- * acts on the match autonomously by sending a DM, so "match" must mean
- * something or AC #5's "no one available" outcome could never be reached).
- */
-export const FIND_HELPER_RELEVANCE_THRESHOLD = KNOWLEDGE_SEARCH_RELEVANCE_THRESHOLD;
-
-export type SetHelperAvailabilityResult = { ok: true } | { ok: false; reason: 'no_interests_row' };
-
-/**
- * Self-scoped flip of the caller's OWN willing_to_help flag. Requires an
- * existing member_interests row (set via set_my_interests) — matching needs
- * that row's published text/embedding to work at all, so there is nothing to
- * flip a flag on for a caller who has never published interests. Instantly
- * reversible either direction, same "one row, caller's own writes replace it"
- * model setMemberInterests uses.
- */
-export async function setHelperAvailability(
-  platform: Platform,
-  userId: string,
-  available: boolean,
-): Promise<SetHelperAvailabilityResult> {
-  const { rowCount } = await pool.query(
-    `UPDATE member_interests SET willing_to_help = $3 WHERE platform = $1 AND user_id = $2`,
-    [platform, userId, available],
-  );
-  return (rowCount ?? 0) > 0 ? { ok: true } : { ok: false, reason: 'no_interests_row' };
-}
-
-/**
- * True if `platform`/`userId` has hit FIND_HELPER_REQUESTER_DAILY_LIMIT
- * successful find_helper handoffs (rows where they're the requester) in the
- * trailing 24h — a call that finds no eligible helper writes no row, so it
- * doesn't count against this cap. Checked BEFORE matching runs (issue #729 AC
- * #6), same DB-backed COUNT(*) shape as createKnowledgeTip's rate cap rather
- * than an in-memory counter.
- */
-export async function isFindHelperRequesterAtDailyCap(platform: Platform, userId: string): Promise<boolean> {
-  const { rows } = await pool.query(
-    `SELECT count(*) AS n FROM helper_notifications
-      WHERE requester_platform = $1 AND requester_user_id = $2
-        AND created_at > now() - interval '24 hours'`,
-    [platform, userId],
-  );
-  return Number(rows[0]?.n ?? 0) >= FIND_HELPER_REQUESTER_DAILY_LIMIT;
-}
-
-export interface HelperCandidate {
-  platform: Platform;
-  userId: string;
-}
-
-/**
- * find_helper's match query — searchMemberInterests' exact shape (embedding-
- * similarity over member_interests, embedding IS NOT NULL), additionally
- * filtered to willing_to_help = true and excluding the requester's own row
- * (SECURITY: issue #729 — self-matching must be impossible even when the
- * requester has willing_to_help = true for their own row). Returns bare
- * candidate identities only, ranked best-first — never the interests text,
- * since the requester's tool result must never see it (that stays
- * who_is_into-only, and even there only for a caller who searches themselves).
- */
-export async function findHelperCandidates(
-  topic: string,
-  excludePlatform: Platform,
-  excludeUserId: string,
-  limit = FIND_HELPER_CANDIDATE_SCAN_LIMIT,
-): Promise<HelperCandidate[]> {
-  let queryVec: number[];
-  try {
-    queryVec = await embed(topic);
-  } catch (err) {
-    logger.warn({ err }, 'Embedding failed for find_helper topic');
-    return [];
-  }
-  const { rows } = await pool.query(
-    `SELECT platform, user_id
-       FROM member_interests
-      WHERE willing_to_help = true AND embedding IS NOT NULL
-        AND NOT (platform = $2 AND user_id = $3)
-        AND 1 - (embedding <=> $1) >= $5
-      ORDER BY embedding <=> $1
-      LIMIT $4`,
-    [pgvector.toSql(queryVec), excludePlatform, excludeUserId, limit, FIND_HELPER_RELEVANCE_THRESHOLD],
-  );
-  return rows.map((r) => ({ platform: r.platform as Platform, userId: r.user_id }));
-}
-
-/**
- * Atomically claims one notification slot for a candidate helper if they're
- * under FIND_HELPER_WEEKLY_LIMIT_PER_HELPER in the trailing 7 days — same
- * `WITH recent AS (...)` restart-proof pattern as createKnowledgeTip, never
- * an in-memory counter (issue #729 SECURITY criterion: pinned by a test that
- * seeds helper_notifications rows directly). Returns false (no row inserted,
- * no DM should be sent) when the helper is already at cap, so the caller can
- * move on to the next candidate; true means this row IS the one-and-only
- * notification for this find_helper call and the caller should now send the DM.
- */
-export async function recordHelperNotificationIfUnderCap(
-  helperPlatform: Platform,
-  helperUserId: string,
-  requesterPlatform: Platform,
-  requesterUserId: string,
-  topic: string,
-): Promise<boolean> {
-  const { rows } = await pool.query(
-    `WITH recent AS (
-       SELECT count(*) AS n FROM helper_notifications
-        WHERE helper_platform = $1 AND helper_user_id = $2
-          AND created_at > now() - interval '7 days'
-     )
-     INSERT INTO helper_notifications
-       (helper_platform, helper_user_id, requester_platform, requester_user_id, topic)
-     SELECT $1, $2, $3, $4, $5
-      WHERE (SELECT n FROM recent) < $6
-     RETURNING id`,
-    [
-      helperPlatform,
-      helperUserId,
-      requesterPlatform,
-      requesterUserId,
-      topic,
-      FIND_HELPER_WEEKLY_LIMIT_PER_HELPER,
-    ],
-  );
-  return rows.length > 0;
 }
 
 // --- Member projects (self-declared project showcase, issue #646) -----------
